Add benchmarks :)
The reference implementation computes the distance between all the words
in a corpus in successive loops.

The initial proposed implementation is based on two cover trees:
the first allows to narrow the coprus to a set of words likely to
appropriate for the wordlist, and the second is used to ensure that
those words are indeed adequate.

The second proposed implementation removes the first cover tree.
It is unlikely that a corpus would be massively oversized, since that
would mean a lot of words in it are increasingly rare.
Using very rare words would defeat the purpose of creating
memoreable passphrases.

Usage:

    go test -bench=.
    # note: please ignore the two benchmarks for inner cover tree functions.